setup issues I was able to begin the setup, and run the backend server and frontend server but my attempts to generate even a basic query result using the react-relay `<QueryRenderer/>` component are producing errors in the browser window.
I have tried a number of solutions involving configuration options, but none seem to produce
the desired result.

Submitting my work so far - I would have liked to get past this to
the actual componentry, but without the graphQL, all I would be able
to show is a static component.